feat: add stub generation for Python bindings
