+++ bindings/python/Cargo.toml
@@ -5,12 +5,18 @@ edition.workspace = true
 
 [lib]
 name = "terminator"
-crate-type = ["cdylib"]
+crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 # PyO3 for Python bindings
-pyo3 = { version = "0.20", features = ["extension-module", "abi3-py311"] }
-pyo3-asyncio = { version = "0.20", features = ["attributes", "tokio-runtime"] }
-terminator = { workspace = true }
+pyo3 = { version = "0.21", features = ["extension-module", "abi3-py311"] }
+pyo3-asyncio-0-21 = { version = "0.21", features = ["attributes", "tokio-runtime"] }
+terminator_core = { path = "../../terminator", package = "terminator" }
 tokio = { workspace = true }
 tracing-subscriber = { workspace = true }
+pyo3-stub-gen = "0.7"
+env_logger = "0.11.8"
+
+[[bin]]
+name = "stub_gen"
+doc = false
