+++ bindings/python/src/bin/stub_gen.rs
@@ -1,0 +1,8 @@
+use pyo3_stub_gen::Result;
+
+fn main() -> Result<()> {
+    env_logger::Builder::from_env(env_logger::Env::default().filter_or("RUST_LOG", "info")).init();
+    let stub = terminator::stub_info()?;
+    stub.generate()?;
+    Ok(())
+}
